+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # CHANGELOG
 
 
+## vN.N.N
+* Update promotion-details-fields to include `slots` schema field.
+
+
 ## v0.6.3
 * Preventing uploader from moving scroll position around on small screens and allowing dynamic height of uploader if viewing area is tiny.
 * Fixing cursor issues with going to line that doesn't exist. Adding support to move cursor to end of last text block.